Implemented 'Spell' component, fixed some CSS issues
