+++ src/Components/App/App.css
@@ -1,10 +1,13 @@
 .Container {
   background-color: #ebd5b3;
+  height: 100%;
+  width: 100%;
 }
 
 .Spellbook {
   display: flex;
   justify-content: center;
+  height: 100%;
 }
 
 .Spellbook-Header {
