+++ src/Components/App/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import cantrips from './CantripsList.js';
-import level1Spells from './Level1List.js';
-import SpellsByLevel from './SpellsByLevel.js';
-import Logo from './images/DnDLogo.jpg'
+import cantrips from '../SpellLists/CantripsList.js';
+import level1Spells from '../SpellLists/Level1List.js';
+import SpellsByLevel from '../SpellsByLevel/SpellsByLevel.js';
+import Logo from '../../images/DnDLogo.jpg'
 import './App.css';
 
 function App() {
@@ -16,7 +16,6 @@ function App() {
         
         <SpellsByLevel spellLevel="Cantrips" spellList={cantrips} />      
         <SpellsByLevel spellLevel="Level 1 Spells" spellList={level1Spells}/>
-        <SpellsByLevel spellLevel="Level 2 Spells" spellList={level1Spells}/>
         
       </div>    
     </div>
