+++ src/Components/SpellLists/Level1List.js
@@ -1,0 +1,29 @@
+const level1Spells = [
+    {
+        spellName: "Acid Splash",
+        spellSchool: "Conjuration",
+        castingTime: "1 Action",
+        range: 60,
+        attackOrSave: "Dex",
+        image: "https://media-waterdeep.cursecdn.com/attachments/2/708/conjuration.png"
+    },
+    {
+        spellName: "Blade Ward",
+        spellSchool: "Abjuration",
+        castingTime: "1 Action",
+        range: "Self",
+        attackOrSave: null,
+        image: "https://encrypted-tbn0.gstatic.com/images?q=tbn%3AANd9GcTBGX-AtXp0Sqbw-pfqwk2rtzE2Il9LohBJdw&usqp=CAU"
+    },
+    {
+        spellName: "Booming Blade",
+        spellSchool: "Evocation",
+        castingTime: "1 Action",
+        range: 5,
+        attackOrSave: "Melee",
+        image: "https://i2.wp.com/dungeonsanddave.com/wp-content/uploads/2017/10/tiago-alexandre-booming-blade-5th-edition-cantrip.jpg?fit=900%2C832&ssl=1&resize=350%2C200"
+    },
+    
+]
+
+export default level1Spells;
