--- src/SpellsByLevel.css
@@ -1,36 +1,0 @@
-.SpellsByLevel {
-    width: 400px;
-    border: black solid;
-    display: inline-block;
-    margin: 5px;
-}
-
-.SpellsByLevel-Header {
-    text-align: center;
-}
-
-.SpellsByLevel-Spell h3 {
-    color: black; 
-    font-family: 'Dancing Script', cursive;
-    font-size: xx-large; 
-    text-decoration: underline;
-    text-align: center;  
-    
-  }
-
-  .SpellsByLevel-Spell ul {
-    list-style-position: inside;
-   }
-
-  .SpellsByLevel-Spell li {
-      text-align: center;    
-  }
-
-  .SpellsByLevel-Spell button {
-    margin: 0 auto;
-    display: block;
-  }
-
-  .Hidden {
-      display: none;
-  }
